+++ extranjeros-web/src/main/java/co/edu/uniandes/csw/extranjeros/dtos/ArrendatarioDetailDTO.java
@@ -18,7 +18,7 @@ public class ArrendatarioDetailDTO extends ArrendatarioDTO {
     // Constructor
     //---------------------------------------------------
     
-    public ArrendatarioDetailDTO(String pUsuario, String pClave, String pCorreo, Long pCelular, String pNombre){
+    public ArrendatarioDetailDTO(String pUsuario, String pClave, String pCorreo, int pCelular, String pNombre){
      super(pUsuario, pClave, pCorreo, pCelular, pNombre);
     }
 }
